pyproject.toml
[tool.poetry]
name = "registrar-reports"
version = "0.1.0"
description = ""
authors = ["tubo <shitubo@gmail.com>"]
readme = "README.md"
packages = [{include = "registrar_reports"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.3"
python-dotenv = "^1.0.0"
playwright = "^1.33.0"
pytest = "^7.3.1"
jupyterlab = "^3.6.3"
pandas = "^2.0.1"
pyparsing = "^3.0.9"
PyYAML = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
